+++ menus.rb
@@ -1,0 +1,32 @@
+class Menus
+  def main_menu
+    puts '1. Books'
+    puts '2. Music Album'
+    puts '3. Games'
+    puts '4. Exit'
+  end
+
+  def books_menu
+    puts 'What would you like to do?'
+    puts '1. List all books'
+    puts '2. List all labels'
+    puts '3. Add a book'
+    puts '4. Go back'
+  end
+
+  def music_menu
+    puts 'What would you like to do?'
+    puts '1. List all music albums'
+    puts '2. List all genres'
+    puts '3. Add a music album'
+    puts '4. Go back'
+  end
+
+  def games_menu
+    puts 'What would you like to do?'
+    puts '1. List of games'
+    puts '2. List all authors'
+    puts '3. Add a game'
+    puts '4. Go back'
+  end
+end
